refactor: move EquationHistory to domain entities and IEquationHistoryRepository to interfaces
