+++ src/presentation.application/domain/equations/interfaces.py
@@ -1,0 +1,5 @@
+from typing import Protocol
+from domain.equations.history.entities import EquationHistory
+
+class IEquationHistoryRepository(Protocol):
+    async def save(self, entity: EquationHistory) -> None: ...
